+++ PrsWebAppSolution/PrsWebAppProject/Controllers/PurchaseRequestLineItemsController.cs
@@ -9,6 +9,7 @@ using PrsWebAppProject.Utility;
 using PrsWebAppProject.Models;
 using Utility;
 using System.Data.Entity;
+using PrsCapstoneProject.Utility;
 
 namespace PrsWebAppProject.Controllers
 {
@@ -16,22 +17,17 @@ namespace PrsWebAppProject.Controllers
     {
         private PrsDbContext db = new PrsDbContext();
 
-        private void UpdatePurchaseRequestTotal(int prodid)
+        private void UpdatePurchaseRequestTotal(int id)
         {
             db = new PrsDbContext();
+            var PurchaseRequest = db.PurchaseRequests.Find(id);
 
-           decimal total = 0;
-           var PurchaseRequestLineItems = db.PurchaseRequestLineItems.Where(p => p.PurchaseRequestId == prodid);
-           foreach (var PurchaseRequestLineItem in PurchaseRequestLineItems)
-           {
-               var Total = PurchaseRequestLineItem.Quantity * PurchaseRequestLineItem.Product.Price;
-           }
-           var PurchaseRequest = db.PurchaseRequests.Find(prodid);
-           PurchaseRequest.Total = total;
+
+            PurchaseRequest.Total = db.PurchaseRequestLineItems.Where(p => p.PurchaseRequestId == PurchaseRequest.Id).Sum(p1 => p1.Product.Price * p1.Quantity);      
            db.SaveChanges();
        }
 
-            //List
+       //List
             public ActionResult List()
         {
             //return Json(db.PurchaseRequestLineItems.ToList(), JsonRequestBehavior.AllowGet);
@@ -44,12 +40,14 @@ namespace PrsWebAppProject.Controllers
         {
             if (id == null)
             {
-                return Json(new JsonMessage("Failure", "Id is null"), JsonRequestBehavior.AllowGet);
+                return new JsonNetResult { Data = new JsonMessage("Failure", "Id is null") };
+                //return Json(new JsonMessage("Failure", "Id is null"), JsonRequestBehavior.AllowGet);
             }
             PurchaseRequestLineItem purchaserequestlineitem = db.PurchaseRequestLineItems.Find(id);
             if (purchaserequestlineitem == null)
             {
-                return Json(new JsonMessage("Failure", "Id is not found"), JsonRequestBehavior.AllowGet);
+                return new JsonNetResult { Data = new JsonMessage("Failure", "Id is not found") };
+                //return Json(new JsonMessage("Failure", "Id is not found"), JsonRequestBehavior.AllowGet);
             }
             return new JsonNetResult { Data = db.PurchaseRequestLineItems.Find(id) };
             //return Json(purchaserequest, JsonRequestBehavior.AllowGet);
@@ -57,27 +55,46 @@ namespace PrsWebAppProject.Controllers
         // Create
         public ActionResult Create([System.Web.Http.FromBody] PurchaseRequestLineItem purchaserequestlineitem)
         {
+            if (purchaserequestlineitem.ProductId == null) return new EmptyResult();
             purchaserequestlineitem.DateCreated = DateTime.Now;
             if (!ModelState.IsValid)
             {
-                return Json(new JsonMessage("Failure", "Model State is not valid"), JsonRequestBehavior.AllowGet);
+                var errorMessages = ModelStateErrors.GetModelStateErrors(ModelState);
+                return new JsonNetResult { Data = new Msg { Result = "Failed", Message = "ModelState invalid.", Data = errorMessages } };
             }
+           
+            
             db.PurchaseRequestLineItems.Add(purchaserequestlineitem);
             try
             {
+               
                 db.SaveChanges();
+
             }
             catch (Exception ex)
             {
-                return Json(new JsonMessage("Failure", ex.Message), JsonRequestBehavior.AllowGet);
+
+                return new JsonNetResult { Data = new JsonMessage("Failure", ex.Message) };
+                //return Json(new JsonMessage("Failure", ex.Message), JsonRequestBehavior.AllowGet);
             }
             UpdatePurchaseRequestTotal(purchaserequestlineitem.PurchaseRequestId);
-            return Json(new JsonMessage("Success", "Purchase Request Line Item was created,"));
+
+
+
+            return new JsonNetResult { Data = new JsonMessage("Success", "Purchase Request Line Item was created", purchaserequestlineitem.Id) };
+
+            //return Json(new JsonMessage("Success", "Purchase Request Line Item was created,"));
         }
 
         //Change
         public ActionResult Change([FromBody] PurchaseRequestLineItem purchaserequestlineitem)
         {
+            if (purchaserequestlineitem.Id == 0) return new EmptyResult();
+            if (!ModelState.IsValid)
+            {
+                var errorMessages = ModelStateErrors.GetModelStateErrors(ModelState);
+                return new JsonNetResult { Data = new Msg { Result = "Failed", Message = "ModelState invalid.", Data = errorMessages } };
+            }
             PurchaseRequestLineItem purchaserequestlineitem2 = db.PurchaseRequestLineItems.Find(purchaserequestlineitem.Id);
             purchaserequestlineitem2.Id = purchaserequestlineitem.Id;
             purchaserequestlineitem2.PurchaseRequestId = purchaserequestlineitem.PurchaseRequestId;
@@ -92,15 +109,24 @@ namespace PrsWebAppProject.Controllers
             }
             catch (Exception ex)
             {
-                return Json(new JsonMessage("Failure", ex.Message), JsonRequestBehavior.AllowGet);
+                return new JsonNetResult { Data = new JsonMessage("Failure", ex.Message) };
+                //return Json(new JsonMessage("Failure", ex.Message), JsonRequestBehavior.AllowGet);
             }
             UpdatePurchaseRequestTotal(purchaserequestlineitem.PurchaseRequestId);
-            return Json(new JsonMessage("Success", "Purchase Request Line Item was changed."));
+            return new JsonNetResult { Data = new JsonMessage("Success", "Purchase Request Line Item was changed") };
+            //return Json(new JsonMessage("Success", "Purchase Request Line Item was changed."));
         }
 
         //Remove
         public ActionResult Remove([FromBody] PurchaseRequestLineItem purchaserequestlineitem)
         {
+            if (purchaserequestlineitem.Id == 0) return new EmptyResult();
+            if (!ModelState.IsValid)
+            {
+                var errorMessages = ModelStateErrors.GetModelStateErrors(ModelState);
+                return new JsonNetResult { Data = new Msg { Result = "Failed", Message = "ModelState invalid.", Data = errorMessages } };
+            }
+            if (purchaserequestlineitem.Id == 0) return new EmptyResult();
             PurchaseRequestLineItem purchaserequestlineitem2 = db.PurchaseRequestLineItems.Find(purchaserequestlineitem.Id);
             db.PurchaseRequestLineItems.Remove(purchaserequestlineitem2);
             try
@@ -109,10 +135,12 @@ namespace PrsWebAppProject.Controllers
             }
             catch (Exception ex)
             {
-                return Json(new JsonMessage("Failure", ex.Message), JsonRequestBehavior.AllowGet);
+                return new JsonNetResult { Data = new JsonMessage("Failure", ex.Message) };
+                //return Json(new JsonMessage("Failure", ex.Message), JsonRequestBehavior.AllowGet);
             }
             UpdatePurchaseRequestTotal(purchaserequestlineitem.PurchaseRequestId);
-            return Json(new JsonMessage("Success", "Purchase Request Line Items were deleted"));
+            return new JsonNetResult { Data = new JsonMessage("Success", "Purchase Request Line Items were deleted") };
+            //return Json(new JsonMessage("Success", "Purchase Request Line Items were deleted"));
         }
     }
